backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daze <dazehere@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
alembic = "^1.5.7"
python-multipart = "^0.0.5"
email-validator = "^1.1.2"
requests = "^2.25.1"
celery = "^5.0.5"
tenacity = "^7.0.0"
pydantic = "^1.8.1"
emails = "^0.6"
raven = "^6.10.0"
gunicorn = "^20.0.4"
Jinja2 = "^2.11.3"
psycopg2-binary = "^2.8.6"
SQLAlchemy = "^1.4.2"
pytest = "^6.2.2"
redis = "^3.5.3"
uvicorn = {version = "^0.13.4", extras = ["standard"]}
google-auth = "^1.28.0"
cryptography = "3.3.1"
python-jose = "^3.2.0"
Faker = "^8.1.0"
Mako = "^1.1.4"
python-dotenv = "^0.17.0"
click = "^7.1.2"
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
black = "^20.8b1"
isort = "^5.8.0"
autoflake = "^1.4"
flake8 = "^3.9.0"
pytest = "^6.2.2"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^2.11.1"
pylint = "^2.7.2"
rope = "^0.18.0"
jedi = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
